+++ resources/views/tailwind/sections/columns.blade.php
@@ -16,7 +16,7 @@
                     class="{{ $column->getVisibility() }} px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider cursor-pointer"
                     id="column_{{ $column->getName() }}"
                     dusk="column-{{ $column->getName() }}"
-                    wire:click="orderBy('{{ $column->getAttribute() }}', '{{ data_get($tableOptions, 'sort.direction') }}')"
+                    wire:click="orderBy('{{ $column->getAttribute() }}', '{{ BelichTables::orderBy($column, $tableOptions) }}')"
                 >
                     <div class="flex justify-start items-center">